refactor: consolidate entity and group
